Create rekognition
In this project, I intend building  image labels generator, using Amazon Rekognition. Once built, it will be able to recognize and label images. For example, if you have a photo of a car, Amazon Recognition will be able to identify what it is, and label the image as a car